Goals for this project
plan.txt has most of what we've been talking about with what we want out
of this game. The multiplayer and death/respawning sections haven't been
discussed yet but uh, read em. And we can discuss.

issues.txt references places in the code I started this summer that are
less than good. Dannenberg expressed interest in refactoring these as a
part of learning how the codebase meshes, and I've expressed interest in
not borking the code further.